Move appearance var helpers to AtomManager Fixes unit tests
